requirements.txt
pyaudio
google-cloud-speech
google-cloud-translate
requests
python-dotenv
webrtcvad
numpy
sounddevice
pyinstaller